pyproject.toml
[project]
name = "ai-saathi-gemini"
version = "0.1.0"
description = "AI Saathi educational platform with Google Gemini integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []